+++ Yerdy/Yerdy/Yerdy_Private.h
@@ -14,6 +14,6 @@
 @property (nonatomic, copy) NSString *ABTag;
 @property (nonatomic, assign) YRDUserType userType;
 
-@property (nonatomic, readonly) int itemsPurchased;
+@property (nonatomic, readonly) NSInteger itemsPurchased;
 
 @end
